Move `BatchPool` to `stateful`
